Add files via upload
Travelling and touring are one of the most popular human activities worldwide. But the choice of the best place of stay after choosing a place to visit becomes of paramount attention.
With the technological advancements across social media and the digitalization of businesses, more and more data have become available across various media platforms.
In the business world today, it paramount for brands to employ digital solutions to drive their businesses forward and the tourist industry is not exempted from this.
Travelling is an essential human endeavour, be it for work or leisure; but getting the right place to stay and the right food, can be a hassle if you have no prior knowledge of the destination.
Following this digital movement, businesses especially hotels and restaurants have now included options for their customers to give feedback for their services in the form of reviews. These reviews become the guide or source of truth for other potential customers who seek to employ their services.
It is now crucial for these data to be explored and utilized in delivering business insights for the public and specifically for business owners.
The study employs the use of text mining and sentiment analysis to evaluate 30 tourist accommodations based on their reviews provided by past users.